+++ src/getContributionIndices.cpp
@@ -4,8 +4,8 @@
 #include <RcppArmadillo.h> //necessary for functions find, intercept
 using namespace arma;
 
-uvec getContributionIndices(int genotypeCombinationIndex, int numberOfContributorIterations,  Mat<uword> outG1allele, int alleleIndexOfDropouts, 
-	int numStutterModels, int *BWto, int *FWto, int numberOfGenotypes, uvec contributionIndices, Col<int> contributorPower ) {
+uvec getContributionIndices(size_t genotypeCombinationIndex, int numberOfContributorIterations,  Mat<uword> outG1allele, int alleleIndexOfDropouts, 
+	int numStutterModels, int *BWto, int *FWto, size_t numberOfGenotypes, uvec contributionIndices, Col<int> contributorPower ) {
 	
 	//Rcpp::Rcout << outG1allele << "-----------\n";  
 	int genoIdx; //used to indicate genotype index
